fix(gatsby): preload and prefetches check for in-browser redirects
